fix typos, since I was reading it anyway
You'll want to review these by hand; I wasn't clear on whether the markdown or the HTML (or something else enriely that I didn't find?) was the original document.  I originally assumed the markdown, so I was editing those files, but then there were some bits that I couldn't find in them, so I switched to the html file.